requirements/project.txt
--requirement=base.txt
--find-links http://downloads.egenix.com/python/index/

-e svn+https://svn.apache.org/repos/asf/incubator/libcloud/trunk#egg=libcloud
-e git+https://github.com/pydanny/django-uni-form.git#egg=django-uni-form
-e git+https://github.com/drewyeaton/django-orienteer.git#egg=django-orienteer
-e git+https://github.com/bfirsh/django-ordered-model.git#egg=django-ordered-model

django_extensions
django-profiles
werkzeug
south
gunicorn
gevent
django-redis-cache
redis
pygeoip
